--- plugins/dev/1756.go
@@ -1,39 +1,0 @@
-package dev
-
-import (
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("1756", "卡尔加里大学", "https://www.ucalgary.ca/")
-
-	engine.SetStartingURLs([]string{"https://www.ucalgary.ca/news/all-ucalgary-news"})
-
-	extractorConfig := extractors.Config{
-		Author:       true,
-		Image:        true,
-		Language:     true,
-		PublishDate:  true,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-
-	engine.OnHTML(".read-more", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Visit(element.Attr("href"), crawlers.News)
-	})
-
-	engine.OnHTML(".col-12.col-lg-8.offset-lg-2.content", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-
-	engine.OnHTML(".pager__item.pager__item--next > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Visit(element.Attr("href"), crawlers.Index)
-	})
-}
