+++ plugins/storage/1754.go
@@ -1,0 +1,42 @@
+package storage
+
+import (
+	"strings"
+
+	"megaCrawler/crawlers"
+	"megaCrawler/extractors"
+
+	"github.com/gocolly/colly/v2"
+)
+
+func init() {
+	engine := crawlers.Register("1754", "罗恩保罗研究所", "https://ronpaulinstitute.org/")
+
+	engine.SetStartingURLs([]string{"https://ronpaulinstitute.org/sitemaps.xml"})
+
+	extractorConfig := extractors.Config{
+		Author:       true,
+		Image:        true,
+		Language:     true,
+		PublishDate:  true,
+		Tags:         true,
+		Text:         false,
+		Title:        true,
+		TextLanguage: "",
+	}
+
+	extractorConfig.Apply(engine)
+
+	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
+		switch {
+		case strings.Contains(element.Text, "post-sitemap"):
+			engine.Visit(element.Text, crawlers.Index)
+		default:
+			engine.Visit(element.Text, crawlers.News)
+		}
+	})
+
+	engine.OnHTML(".et_pb_with_border.et_pb_column_3_4", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
+	})
+}
