+++ plugins/storage/1755.go
@@ -1,0 +1,42 @@
+package storage
+
+import (
+	"strings"
+
+	"megaCrawler/crawlers"
+	"megaCrawler/extractors"
+
+	"github.com/gocolly/colly/v2"
+)
+
+func init() {
+	engine := crawlers.Register("1755", "美国达特茅斯学院", "https://home.dartmouth.edu/")
+
+	engine.SetStartingURLs([]string{"https://home.dartmouth.edu/sitemap.xml"})
+
+	extractorConfig := extractors.Config{
+		Author:       true,
+		Image:        true,
+		Language:     true,
+		PublishDate:  true,
+		Tags:         true,
+		Text:         false,
+		Title:        true,
+		TextLanguage: "",
+	}
+
+	extractorConfig.Apply(engine)
+
+	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
+		if strings.Contains(element.Text, "sitemap") {
+			engine.Visit(element.Text, crawlers.Index)
+		}
+		if strings.Contains(element.Text, "/news/") {
+			engine.Visit(element.Text, crawlers.News)
+		}
+	})
+
+	engine.OnHTML(".field__item", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
+	})
+}
