+++ plugins/storage/1758.go
@@ -1,0 +1,39 @@
+package storage
+
+import (
+	"strings"
+
+	"megaCrawler/crawlers"
+	"megaCrawler/extractors"
+
+	"github.com/gocolly/colly/v2"
+)
+
+func init() {
+	engine := crawlers.Register("1758", "新南威尔士大学", "https://www.unsw.edu.au/")
+
+	engine.SetStartingURLs([]string{"https://www.unsw.edu.au/sitemap.xml"})
+
+	extractorConfig := extractors.Config{
+		Author:       true,
+		Image:        true,
+		Language:     true,
+		PublishDate:  true,
+		Tags:         true,
+		Text:         false,
+		Title:        true,
+		TextLanguage: "",
+	}
+
+	extractorConfig.Apply(engine)
+
+	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
+		if strings.Contains(element.Text, "/newsroom/news/20") {
+			ctx.Visit(element.Text, crawlers.News)
+		}
+	})
+
+	engine.OnHTML(".cmp-text", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
+	})
+}
