+++ plugins/storage/1759.go
@@ -1,0 +1,39 @@
+package storage
+
+import (
+	"megaCrawler/crawlers"
+	"megaCrawler/extractors"
+
+	"github.com/gocolly/colly/v2"
+)
+
+func init() {
+	engine := crawlers.Register("1759", "沃尔特·E·华盛顿会议中心", "https://eventsdc.com/")
+
+	engine.SetStartingURLs([]string{"https://eventsdc.com/node"})
+
+	extractorConfig := extractors.Config{
+		Author:       true,
+		Image:        true,
+		Language:     true,
+		PublishDate:  true,
+		Tags:         true,
+		Text:         false,
+		Title:        true,
+		TextLanguage: "",
+	}
+
+	extractorConfig.Apply(engine)
+
+	engine.OnHTML(".node > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Visit(element.Attr("href"), crawlers.News)
+	})
+
+	engine.OnHTML(".field.field--name-body", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
+	})
+
+	engine.OnHTML(".pager__item.pager__item--next > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Visit(element.Attr("href"), crawlers.Index)
+	})
+}
